+++ src/dashboardWebView/components/Media/Item.tsx
@@ -1,6 +1,6 @@
 import { Messenger } from '@estruyf/vscode/dist/client';
 import { CheckCircleIcon, ClipboardCopyIcon, CodeIcon, PencilIcon, PhotographIcon, TrashIcon } from '@heroicons/react/outline';
-import { basename, dirname } from 'path';
+import { basename, dirname, parse } from 'path';
 import * as React from 'react';
 import { useEffect } from 'react';
 import { useRecoilState, useRecoilValue } from 'recoil';
@@ -23,6 +23,7 @@ export const Item: React.FunctionComponent<IItemProps> = ({media}: React.PropsWi
   const viewData = useRecoilValue(ViewDataSelector);
   const [ caption, setCaption ] = React.useState(media.caption);
   const [ alt, setAlt ] = React.useState(media.alt);
+  const [ filename, setFilename ] = React.useState<string | null>(null);
   const page = useRecoilValue(PageSelector);
 
   const parseWinPath = (path: string | undefined) => {
@@ -99,14 +100,24 @@ export const Item: React.FunctionComponent<IItemProps> = ({media}: React.PropsWi
   };
 
   const calculateSize = () => {
+    let sizeDetails = [];
+
+    if (media?.dimensions) {
+      if (media.dimensions.width && media.dimensions.height) {
+        sizeDetails.push(`${media.dimensions.width}x${media.dimensions.height}`);
+      }
+    }
+
     if (media?.stats?.size) {
       const size = media.stats.size / (1024*1024);
       if (size > 1) {
-        return `${size.toFixed(2)} MB`;
+        sizeDetails.push(`${size.toFixed(2)} MB`);
       } else {
-        return `${(size * 1024).toFixed(2)} KB`;
+        sizeDetails.push(`${(size * 1024).toFixed(2)} KB`);
       }
     }
+
+    return sizeDetails.join(" — ");
   };
 
   const openLightbox = () => {
@@ -120,6 +131,7 @@ export const Item: React.FunctionComponent<IItemProps> = ({media}: React.PropsWi
   const submitMetadata = () => {
     Messenger.send(DashboardMessage.updateMediaMetadata, {
       file: media.fsPath,
+      filename,
       caption,
       alt,
       folder: selectedFolder,
@@ -139,6 +151,17 @@ export const Item: React.FunctionComponent<IItemProps> = ({media}: React.PropsWi
       setCaption(media.caption);
     }
   }, [media.caption]);
+  
+  useEffect(() => {
+    const name = basename(parseWinPath(media.fsPath) || "");
+    if (name !== filename) {
+      setFilename(name);
+    }
+  }, [media.fsPath]);
+
+  const fileInfo = filename ? basename(filename).split('.') : null;
+  const extension = fileInfo?.pop();
+  const name = fileInfo?.join('.');
 
   return (
     <>
@@ -151,8 +174,15 @@ export const Item: React.FunctionComponent<IItemProps> = ({media}: React.PropsWi
             <img src={media.vsPath} alt={basename(media.fsPath)} className="mx-auto object-cover" />
           </div>
         </button>
-        <div className={`relative py-4 pl-4 pr-10`}>
+        <div className={`relative py-4 pl-4 pr-12`}>
           <div className={`absolute top-4 right-4 flex flex-col space-y-4`}>
+            <button title={`Edit metadata`} 
+                    className={`hover:text-teal-900 focus:outline-none`} 
+                    onClick={updateMetadata}>
+              <PencilIcon className={`h-5 w-5`} />
+              <span className={`sr-only`}>Edit metadata</span>
+            </button>
+
             {
               viewData?.data?.filePath ? (
                 <>
@@ -177,12 +207,6 @@ export const Item: React.FunctionComponent<IItemProps> = ({media}: React.PropsWi
                 </>
               ) : (
                 <>
-                  <button title={`Edit metadata`} 
-                          className={`hover:text-teal-900 focus:outline-none`} 
-                          onClick={updateMetadata}>
-                    <PencilIcon className={`h-5 w-5`} />
-                    <span className={`sr-only`}>Edit metadata</span>
-                  </button>
                   <button title={`Copy media path`} 
                           className={`hover:text-teal-900 focus:outline-none`} 
                           onClick={copyToClipboard}>
@@ -223,7 +247,7 @@ export const Item: React.FunctionComponent<IItemProps> = ({media}: React.PropsWi
             <span className={`block mt-1 dark:text-whisper-500 text-xs`}>{getFolder()}</span>
           </p>
           {
-            media?.stats?.size && (
+            (media?.stats?.size || media?.dimensions) && (
               <p className="mt-2 text-xs dark:text-whisper-900 font-medium pointer-events-none flex flex-col items-start">
                 <b className={`mr-1`}>Size:</b>
                 <span className={`block mt-1 dark:text-whisper-500 text-xs`}>{calculateSize()}</span>
@@ -241,8 +265,26 @@ export const Item: React.FunctionComponent<IItemProps> = ({media}: React.PropsWi
             okBtnText={`Save`}
             cancelBtnText={`Cancel`}
             dismiss={() => setShowForm(false)}
-            trigger={submitMetadata}>
+            trigger={submitMetadata}
+            isSaveDisabled={!filename}>
             <div className="flex flex-col space-y-2">
+              <div>
+                <label htmlFor="about" className="block text-sm font-medium text-gray-700 dark:text-whisper-900">
+                  Filename
+                </label>
+                <div className="relative mt-1">
+                  <input
+                    className="py-1 px-2 sm:text-sm bg-white dark:bg-vulcan-300 border border-gray-300 dark:border-vulcan-100 text-vulcan-500 dark:text-whisper-500 placeholder-gray-400 dark:placeholder-whisper-800 focus:outline-none w-full"
+                    value={name}
+                    onChange={(e) => setFilename(`${e.target.value}.${extension}`)} />
+
+                  <div className="absolute inset-y-0 right-0 pr-3 flex items-center pointer-events-none">
+                    <span className="text-gray-500 sm:text-sm">
+                      .{extension}
+                    </span>
+                  </div>
+                </div>
+              </div>
               <div>
                 <label htmlFor="about" className="block text-sm font-medium text-gray-700 dark:text-whisper-900">
                   Caption