+++ src/dashboardWebView/components/Modals/Metadata.tsx
@@ -7,12 +7,13 @@ export interface IMetadataProps {
   description: string;
   okBtnText: string;
   cancelBtnText: string;
+  isSaveDisabled: boolean;
 
   dismiss: () => void;
   trigger: () => void;
 }
 
-export const Metadata: React.FunctionComponent<IMetadataProps> = ({title, description, cancelBtnText, okBtnText, dismiss, trigger, children}: React.PropsWithChildren<IMetadataProps>) => {
+export const Metadata: React.FunctionComponent<IMetadataProps> = ({title, description, cancelBtnText, okBtnText, dismiss, isSaveDisabled, trigger, children}: React.PropsWithChildren<IMetadataProps>) => {
 
   const cancelButtonRef = useRef(null);
   
@@ -66,8 +67,9 @@ export const Metadata: React.FunctionComponent<IMetadataProps> = ({title, descri
                 <div className="mt-5 sm:mt-4 sm:flex sm:flex-row-reverse">
                   <button
                     type="button"
-                    className="w-full inline-flex justify-center rounded-md border border-transparent shadow-sm px-4 py-2 bg-teal-600 text-base font-medium text-white hover:bg-teal-700 dark:hover:bg-teal-900 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-teal-500 sm:ml-3 sm:w-auto sm:text-sm"
+                    className="w-full inline-flex justify-center rounded-md border border-transparent shadow-sm px-4 py-2 bg-teal-600 text-base font-medium text-white hover:bg-teal-700 dark:hover:bg-teal-900 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-teal-500 sm:ml-3 sm:w-auto sm:text-sm disabled:opacity-30"
                     onClick={() => trigger()}
+                    disabled={isSaveDisabled}
                   >
                     {okBtnText}
                   </button>