+++ src/models/MediaPaths.ts
@@ -1,4 +1,5 @@
 import { Stats } from "fs";
+import { ISizeCalculationResult } from "image-size/dist/types/interface";
 
 export interface MediaPaths { 
   media: MediaInfo[];
@@ -10,6 +11,7 @@ export interface MediaInfo {
   fsPath: string; 
   vsPath: string | undefined;
   stats: Stats | undefined;
+  dimensions: ISizeCalculationResult | undefined;
   caption?: string | undefined;
   alt?: string | undefined;
 }